Restructure FS726T.queue_change
Instead of modifying the changes list in-place, we now create a new list
and replace it at the end. This simplifies the code without having to
worry about modifying the same list we're looping over. This also fixes
a subtle bug where the old code would skip over an element if the
previous element was removed.