Log out when quitting
These switches are picky with logins, disallowing logins from other IP
addresses when logged in. To prevent issues, it is important to actually
log out when you are done, so you can later log in from another IP
address.

There was already code to log out, but it was not being called. This
commit calls the logout code when quitting and modifies the code to no
automatically log back in after logging out (resulting in an infinite
logout/login loop).