Add workaround for newer urwid versions
It seems in newer urwid versions, Pile and Columns cache their
selectable attribute, which breaks with the PortVlanMatrix that would
only become selectable once the switch status is parsed and it contains
VLANs.

As a workaround, make PortVlanMatrix always selectable, which causes the
containing Columns and Pile objects to do the same.