Move vlan name handling into FS726T class
It seems most other switches can just store VLAN names in the switch
itself, so storing them in the config is just a limitation of this
particular switch.

This change does mean that the config is now written for every name
change (or vlan delete) instead of once after the entire commit, but
that should be fine.