Support more ports than fit on screen
This inverts the internal structure of the PortVlanMatrix. Before it was
a Pile of Columns, which caused just a single row to scroll when the
focus when past the end of the screen. Now it is a Columns of Piles,
which means all columns scroll together.

This does mean that when moving the cursor horizontally, it jumps back
to the first row every time (since Columns does not preserve vertical
cursor position like Pile does for horizontal cursor position), but this
is fixed by propagating the vertical focus to all columns on every
keypress.